backend/requirements.txt
flask
Flask-RESTful
requests
python-dotenv
langchain
langchain-core
langchain-chroma
langchain-community
openai